chore: switch to Github workers
