+++ log-app/wrangler.toml
@@ -1,0 +1,3 @@
+name = "log-app"
+main = "src/index.ts"
+compatibility_date = "2023-08-19"
