--- log-webapp/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "cf-pages-functions",
-  "version": "0.0.0",
-  "scripts": {
-    "build": "tsc"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^3.2.0",
-    "@supabase/supabase-js": "^2.32.0",
-    "typescript": "^4.3.2"
-  },
-  "dependencies": {}
-}
