--- log-webapp/types/global.ts
@@ -1,5 +1,0 @@
-export interface Env {
-  ASSETS: Fetcher;
-  SUPABASE_KEY: string;
-  SUPABASE_URL: string;
-}
